v1/functions/pages/requirements.txt
orator
PyMySQL
PyJWT
python-slugify
